+++ game-core/src/main/java/games/strategy/engine/stats/Statistics.java
@@ -13,4 +13,5 @@ import lombok.Value;
 @Value
 public class Statistics {
   private final Table<String, Round, Double> productionOfPlayerInRound = HashBasedTable.create();
+  private final Table<String, Round, Double> tuvOfPlayerInRound = HashBasedTable.create();
 }
